@@ -12,7 +12,6 @@
     "lint:safe": "eslint ./src --ext=js,vue ;exit 0",
     "build": "webpack --progress --hide-modules --config webpack.config.js",
     "serve": "webpack-dev-server --hot --inline --config webpack.config.js -d",
-    "deploy": "npm run build;gulp deploy",
     "test": "npm run lint",
     "test:safe": "npm run lint; exit 0"
   },
@@ -26,25 +25,26 @@
   },
   "homepage": "https://github.com/potato4d/vue-dev-template#readme",
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.16.0",
+    "axios": "^0.17.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
     "css-loader": "^0.26.2",
     "dotenv": "^4.0.0",
     "dotenv-webpack": "^1.4.2",
     "eslint": "^3.8.1",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
     "eslint-plugin-vue": "^2.0.1",
-    "gulp": "^3.9.1",
-    "gulp-s3-upload": "^1.6.1",
-    "vue": "^2.0.3",
-    "vue-loader": "^11.1.4",
-    "vue-router": "^2.0.1",
+    "vue": "~2.5.2",
+    "vue-loader": "^13.0.5",
+    "vue-router": "^2.7.0",
+    "vue-server-renderer": "~2.5.2",
     "vue-style-loader": "^2.0.3",
-    "vue-template-compiler": "^2.2.1",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1",
-    "whatwg-fetch": "^2.0.3"
+    "vue-template-compiler": "~2.5.2",
+    "webpack": "^3.6.0",
+    "webpack-dev-server": "^2.9.3"
   }
 }
