@@ -1,7 +1,7 @@
 <template lang="html">
-	<p>
-		Hello, \{{name}}
-	</p>
+  <p>
+    Hello, \{{name}}
+  </p>
 </template>
 
 <style scoped>
@@ -9,11 +9,11 @@
 </style>
 
 <script>
-module.exports = {
-	props: ["name"],
-	data: () => {
-		return {
-		}
-	}
+export default {
+  props: ['name'],
+  data () {
+    return {
+    }
+  }
 }
 </script>
