@@ -1,7 +1,7 @@
 <template lang="html">
-	<div class="content">
-		<greeting name="Vue.js"></greeting>
-	</div>
+  <div class="content">
+    <greeting name="Vue.js"></greeting>
+  </div>
 </template>
 
 <style scoped>
@@ -9,11 +9,17 @@
 </style>
 
 <script>
-module.exports = {
-	data: () => {
-		return {
-			stores: require("../../stores/Stores.js")
-		};
-	}
+import Greeting from '@/components/Greeting.vue'
+import Stores from '@/stores/Stores'
+
+export default {
+  components: {
+    Greeting
+  },
+  data () {
+    return {
+      Stores
+    }
+  }
 }
 </script>
