@@ -1,3 +1,3 @@
-module.exports = {
-  ApplicationStore: require("./ApplicationStore")
+export default {
+
 }
