@@ -1,35 +1,9 @@
 module.exports = {
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es6": true,
-        "node": true
-    },
-	"globals": {
-		"ga":false
-	},
-    "extends": "eslint:recommended",
-    "rules": {
-        "indent": [
-            "error",
-            "tab"
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "quotes": [
-            "error",
-            "double"
-        ],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "no-console": "off",
-        "no-var": "error"
-    },
-    "plugins": [
-        "vue"
-    ],
+  extends: 'standard',
+  "env": {
+    "browser": true,
+  },
+  "plugins": [
+    "vue"
+  ],
 };
